chore(logger): expose logger._buffer={}, add tests
This commit exposes the private buffers publicly
_buffer={events=[],metrics=[],tracking=[]} on the logger instance.

In `3.x` this was exposed on `$beaver` but was removed in later
versions.

Adding this allows us much greater flexibility in testing as well as
allows consuming libraries to assert that they sent the correct events
to the buffer for logging.

This change also allows us to add a lot more tests easily (code coverage
jumped to 73%).